# api/swagger/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/brands-list": {
      "get": {
        "operationId": "BrandListController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name or email, case-insensitive",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EAccountStatus"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EAccountSortField"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ESortOrder"
            }
          },
          {
            "name": "emailVerified",
            "required": false,
            "in": "query",
            "description": "true = email đã xác minh, false = chưa",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "createdFrom",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "createdTo",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-12-31",
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": false,
            "in": "query",
            "description": "Match against the address in the role profile table",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountFilterResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Requires admin role"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List brand accounts, paginated",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/brands-list/{id}": {
      "get": {
        "operationId": "BrandListController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired\n\nToken is missing, invalid or expired"
          },
          "403": {
            "description": "Requires admin role\n\nRequires admin role"
          },
          "404": {
            "description": "Brand profile not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Brand account details",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/admin/admin-list": {
      "get": {
        "operationId": "AdminController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by name or email, case-insensitive",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by account status",
            "schema": {
              "$ref": "#/components/schemas/EAccountStatus"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EAccountSortField"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ESortOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFilterResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Requires admin role"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List admin accounts, paginated",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/{id}/status": {
      "patch": {
        "operationId": "AdminController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Requires admin role"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change account status; banning takes effect immediately",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/{id}": {
      "get": {
        "operationId": "AdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Account exists but is not an admin"
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Requires admin role"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Read one admin account in detail",
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "operationId": "AdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body, malformed IP/CIDR, or account is not an admin"
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Requires super admin role, or IP not whitelisted"
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Email already belongs to another account"
          },
          "422": {
            "description": "Change would lock the caller out of their own whitelist; set acknowledgeSelfLockout to override"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update one admin account. Super admin only",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Account exists but is not an admin"
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Requires super admin role, or IP not whitelisted"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete one admin account. Super admin only",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM KOC API",
    "description": "REST API for the CRM KOC system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ERole": {
        "type": "string",
        "enum": [
          "admin",
          "brand",
          "creator"
        ]
      },
      "EAccountStatus": {
        "type": "number",
        "enum": [
          1,
          2,
          3,
          4
        ]
      },
...
      "EAccountSortField": {
        "type": "string",
        "enum": [
          "createdAt",
          "updatedAt",
          "name",
          "email",
          "status"
        ]
      },
      "ESortOrder": {
        "type": "string",
        "enum": [
          "ASC",
          "DESC"
        ]
      },
      "AccountFilterItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+84901111111"
          },
          "accountRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ERole"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EAccountStatus"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "accountRole",
          "status",
          "createdAt"
        ]
      },
      "AccountFilterResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountFilterItemDto"
            }
          },
          "total": {
            "type": "number",
            "example": 3,
            "description": "Total rows matching the filter"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "EAdminRole": {
        "type": "string",
        "enum": [
          "admin",
          "super_admin"
        ]
      },
      "AdminResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ERole"
              }
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EAccountStatus"
              }
            ]
          },
          "statusReason": {
            "type": "string",
            "nullable": true
          },
          "emailVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "phoneVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "ipWhitelist": {
            "type": "string",
            "nullable": true,
            "example": "203.0.113.1,10.0.0.0/24",
            "description": "null hoặc rỗng = cho phép truy cập từ mọi IP"
          },
          "adminRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EAdminRole"
              }
            ]
          }
        },
        "required": [
          "id",
          "accountRole",
          "name",
          "email",
          "phone",
          "status",
          "statusReason",
          "emailVerifiedAt",
          "phoneVerifiedAt",
          "createdAt",
          "updatedAt",
          "ipWhitelist",
          "adminRole"
        ]
      },
      "AdminFilterResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total rows matching the filter"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "UpdateStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EAccountStatus"
              }
            ]
          },
          "statusReason": {
            "type": "object",
            "example": "Spamming campaigns",
            "description": "Shown to the account owner when they are blocked"
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "example": "admin@gmail.com",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 20
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EAccountStatus"
              }
            ]
          },
          "statusReason": {
            "type": "object",
            "example": "Spamming campaigns",
            "description": "Shown to the account owner when they are blocked"
          },
          "ipWhitelist": {
            "type": "string",
            "example": "203.0.113.1,10.0.0.0/24",
            "maxLength": 2000,
            "description": "CSV các IP/CIDR. Ghi đè toàn bộ. Rỗng = cho phép mọi IP."
          },
          "acknowledgeSelfLockout": {
            "type": "boolean",
            "default": false,
            "description": "Bỏ qua kiểm tra tự khoá khi tự sửa whitelist của mình."
          }
        }
      },
...
    }
  }
}